scheme: add new scheme move to scheme directory
