+++ scheme/sui/sui.go
@@ -1,0 +1,1 @@
+package sui
