+++ types/scheme.go
@@ -5,4 +5,6 @@ type Scheme string
 const (
 	EVM    Scheme = "evm"
 	Solana Scheme = "solana"
+	Sui    Scheme = "sui"
+	Tron   Scheme = "tron"
 )
